TIP-1403: move namespace
